pyproject.toml
[project]
name = "bomshell"
version = "2.0.0"
description = "Retrieve weather data from the Australian Bureau of Meteorology"
readme = "README.rst"
license = "GPL-2.0-only"
authors = [{ name = "sthysel", email = "sthysel@gmail.com" }]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "requests>=2.28",
    "pyxdg>=0.28",
    "dbfread>=2.0",
    "dataset>=1.6",
    "ftputil>=5.0",
    "metar>=1.11",
    "python-dotenv>=1.0",
    "folium>=0.15",
    "geopandas>=0.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "ty>=0.0.1a7",
    "pre-commit>=3.7",
    "bump-my-version>=0.20",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
]

[project.scripts]
bomshell = "bomshell.cli:main"

[project.urls]
Homepage = "https://github.com/sthysel/bomshell"
Repository = "https://github.com/sthysel/bomshell"
Documentation = "https://bomshell.readthedocs.io"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bomshell"]

[tool.ruff]
target-version = "py310"
line-length = 140
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["bomshell"]
force-single-line = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-rxEfsw --strict-markers --tb=short"

[tool.coverage.run]
source = ["src/bomshell"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.bumpversion]
current_version = "2.0.0"
commit = true
tag = true
tag_name = "v{new_version}"
tag_message = "Release {new_version}"
message = "Bump version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/bomshell/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "docs/conf.py"
search = 'version = release = "{current_version}"'
replace = 'version = release = "{new_version}"'
